validacion mediante token en el modulo de productoController

+++ ProyectoFinal/controllers/productoController.js
@@ -20,14 +20,14 @@ const agregarProductoGET = async (req, res) => {
                 //     titulo: "Agregar producto"
                 // });
             } else {
-                return res.status(401).send({error: 'ERROR TOKEN EXPIRADO O NO VALIDO'});
+                return res.status(401).send({error: 'Error token expirado o no valido'});
             }
         } else {
             // res.redirect('/login');
-            return res.status(401).send({ error: 'DEBE INICIAR SECION' })
+            return res.status(401).send({ error: 'Debe iniciar sesion' });
         }
     } catch (error) {
-        res.status(500).send({ error: 'ERROR NO TIENE AUTORIZACION' })
+        res.status(500).send({ error: 'ERROR NO TIENE AUTORIZACION' });
         console.log('[ERROR]' + error);
     };
 };
@@ -36,24 +36,35 @@ const agregarProductoPOST = async (req, res) => {
 
     try {
         if(req.session.logueado === true) {
-            const nuevoProducto = await Productos.create({
-                nombre: req.body.nombre,
-                descripcion: req.body.descripcion,
-                caracteristica: req.body.caracteristica,
-                precio: req.body.precio,
-                stock: req.body.stock,
-                rutaImagen: req.body.rutaImagen,
-                destacado: req.body.destacado
-            })
-            console.log(nuevoProducto);
-            res.render("agregar-producto", { 
-                mensaje: "Producto agregado",
-                titulo: "Agregar producto"
-            });
+
+            const token = req.headers.authorization.split(' ')[1];
+            const tokenData = await tokenVerify(token);
+
+            if(Date.now() < tokenData.expiresIn){
+                const nuevoProducto = await Productos.create({
+                    nombre: req.body.nombre,
+                    descripcion: req.body.descripcion,
+                    caracteristica: req.body.caracteristica,
+                    precio: req.body.precio,
+                    stock: req.body.stock,
+                    rutaImagen: req.body.rutaImagen,
+                    destacado: req.body.destacado
+                });
+                res.status(200).send({
+                    titulo: 'Agregar Producto',
+                    message: 'Producto agregado correctamente',
+                    producto: nuevoProducto,
+                });
+                // res.render("agregar-producto", { 
+            } else {
+                return res.status(401).send({error: 'Error token expirado o no valido'});
+            };
         } else {
-            res.redirect('/login');
+            res.status(401).send({ error: 'Debe iniciar sesion' });
+            // res.redirect('/login');
         };
     } catch (error) {
+        res.status(500).send({ error: 'ERROR NO TIENE AUTORIZACION' });
         console.log('[ERROR]' + error);
     };    
 };
@@ -63,28 +74,42 @@ const editarProductoGET = async (req, res) => {
 
     try {
         if(req.session.logueado === true){
-            const productoFind = await Productos.findOne({ 
-                where: {
-                    id: idProducto
-                } 
-            })
-    
-            if( !productoFind ) {
-                res.send(`
-                    <h1>No existe el producto con id ${idProducto}</h1>
-                    <a href="/admin"> Ver listado de productos</a>
-                `)
+
+            const token = req.headers.authorization.split(' ')[1];
+            const tokenData = await tokenVerify(token);
+
+            if(Date.now() < tokenData.expiresIn){
+
+                const productoFind = await Productos.findOne({ 
+                    where: {
+                        id: idProducto
+                    }
+                });
+                if( !productoFind ) {
+                    res.send(`
+                        <h1>No existe el producto con id ${idProducto}</h1>
+                        <a href="/admin"> Ver listado de productos</a>
+                    `);
+                } else {
+                    res.status(200).send({
+                        titulo: "Vista del Administrador, editar",
+                        productos: productoFind
+                    });
+                    // res.render('editar-producto', {
+                    //     titulo: "Vista del administrador",
+                    //     productos: productoFind
+                    // });    
+                };
             } else {
-                res.render('editar-producto', {
-                    titulo: "Vista del administrador",
-                    productos: productoFind
-                });    
-            };
+                return res.status(401).send({ error: 'Ddebe iniciar sesion' });
+            }
         } else {
-            res.redirect('/login');
+            return res.status(401).send({ error: 'Error en la verificacion o token no valido' });
+            // res.redirect('/login');
         }
     } catch (error) {
-        console.log(error);
+        res.status(500).send({ error: 'ERROR NO TIENE AUTORIZACION' });
+        console.log('[ERROR]' + error);
     };
 };
 
@@ -94,47 +119,69 @@ const editarProductoPOST = async (req, res) => {
     
     try {
         if( req.session.logueado === true ){
-            const nuevoProducto = await Productos.update({
-                nombre: producto.nombre,
-                descripcion: producto.descripcion,
-                caracteristica: producto.caracteristica,
-                precio: producto.precio,
-                stock: producto.stock,
-                rutaImagen: producto.rutaImagen,
-                destacado: producto.destacado
-            }, 
-            {
-                where: { id: idProducto }
-            })
-            console.log('SE MODIFICO CORRECTAMENTE')
-            // console.log(nuevoProducto);
-            res.redirect('/admin');
+
+            const token = req.headers.authorization.split(' ')[1];
+            const tokenData = await tokenVerify(token);
+
+            if(Date.now() < tokenData.expiresIn){
+                const nuevoProducto = await Productos.update({
+                    nombre: producto.nombre,
+                    descripcion: producto.descripcion,
+                    caracteristica: producto.caracteristica,
+                    precio: producto.precio,
+                    stock: producto.stock,
+                    rutaImagen: producto.rutaImagen,
+                    destacado: producto.destacado
+                }, 
+                {
+                    where: { id: idProducto }
+                });
+                res.status(200).send({ 
+                    message: 'Producto modificado correctamente'
+                });
+                // res.redirect('/admin');
+            } else {
+                res.status(401).send({ error: 'Errot token invalido o expirado'});
+            }
         } else {
-            res.redirect('/login');
+            res.status(401).send({ error: 'Debe iniciar sesion'});
+            // res.redirect('/login');
         };
     } catch (error) {
+        res.status(500).send({ error: 'ERROR NO AUTORIZADO'});
         console.log('[ERROR]' + error);
     };
 };
 
 const borrarProductoGET = async (req, res) => {
-    const idProducto = req.params.id
+    const idProducto = req.params.id;
 
     try {
         if( req.session.logueado === true ) {
-            const productoDelete = await Productos.destroy({
-                where: {
-                    id: idProducto
-                }
-            });
-            console.log('PRODUCTO ELIMINADO');
-            // res.send('Producto eliminado')
-            res.redirect('/admin');
+
+            const token = req.headers.authorization.split(' ')[1];
+            const tokenData = await tokenVerify(token);
+
+            if(Date.now() < tokenData.expiresIn){
+                const productoDelete = await Productos.destroy({
+                    where: {
+                        id: idProducto
+                    }
+                });
+                console.log('*** PRODUCTO ELIMINADO ***');
+                return res.status(200).send({ message: 'producto eliminado correctamente'});
+                // res.redirect('/admin');
+
+            } else {
+                return res.status(401).send({error: 'Error token expirado o no validado'});
+            }
         } else {
-            res.redirect('/login');
+            return res.status(401).send({error: 'Debe iniciar sesion'});
+            // res.redirect('/login');
         }
     } catch (error) {
         console.log('[ERROR]' + error);
+        res.status(500).send({ error: 'ERROR NO TIENE AUTORIZACION' });
     };    
 };
 
@@ -146,4 +193,4 @@ module.exports = {
     editarProductoGET,
     editarProductoPOST,
     borrarProductoGET,
-}
+};
